requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.1.2
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2.6
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
h11==0.12.0
hiredis==2.0.0
httptools==0.2.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
iniconfig==1.1.1
kombu==5.1.0
msgpack==1.0.2
packaging==21.0
pluggy==0.13.1
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.4
python-crontab==2.5.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.0
uvicorn==0.15.0
uvloop==0.16.0
vine==5.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==9.1
zope.interface==5.4.0
